refactor: permitAll() 문제 해결
