+++ src/main/java/com/circle/circle_backend/security/config/SecurityConfig.java
@@ -5,9 +5,7 @@ import com.circle.circle_backend.security.filter.JwtAuthenticationFilter;
 import com.circle.circle_backend.security.filter.JwtAuthorizationFilter;
 import com.circle.circle_backend.security.service.UserDetailsServiceImpl;
 import com.circle.circle_backend.security.utils.JwtTokenUtils;
-import jakarta.servlet.Filter;
 import lombok.RequiredArgsConstructor;
-import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
@@ -15,6 +13,7 @@ import org.springframework.security.authentication.ProviderManager;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.factory.PasswordEncoderFactories;
@@ -31,6 +30,12 @@ public class SecurityConfig {
     private final JwtTokenUtils jwtTokenUtils;
     private final UserDetailsServiceImpl userDetailsServiceImpl;
     private final CorsFilter corsFilter;
+
+    @Bean
+    public WebSecurityCustomizer webSecurityCustomizer() {
+        return web -> web.ignoring()
+                    .requestMatchers("/api/login", "/api/users");
+    }
 
     @Bean
     public PasswordEncoder passwordEncoder() {
@@ -54,7 +59,7 @@ public class SecurityConfig {
     }
 
     @Bean
-    public Filter jwtAuthorizationFilter() {
+    public JwtAuthorizationFilter jwtAuthorizationFilter() {
         return new JwtAuthorizationFilter(jwtTokenUtils, userDetailsServiceImpl);
     }
 
@@ -68,16 +73,15 @@ public class SecurityConfig {
                         sessionManagement.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
                 )
 
-                // 요정 인증 설정
-                .authorizeHttpRequests((authorizeHttpRequests) -> authorizeHttpRequests
-                        .requestMatchers(PathRequest.toStaticResources().atCommonLocations()).permitAll() // resources 접근 허용 설정
-                        .requestMatchers("/api/**").permitAll() // `/api/users`는 인증 없이 접근 가능
-                        .anyRequest().authenticated() // 그 외의 모든 요청은 인증 필요
+                // 요청 인증 설정
+                .authorizeHttpRequests(authorizeHttpRequests -> authorizeHttpRequests
+                        .anyRequest().authenticated() // 그 외 요청은 인증 필요
                 )
                 // 필터 순서 설정
-                .addFilterBefore(corsFilter, ChannelProcessingFilter.class) // CORS 필터를 가장 앞단에 배치
-                .addFilterBefore(jwtAuthenticationFilter(), JwtAuthorizationFilter.class)
+                .addFilterBefore(corsFilter, ChannelProcessingFilter.class) // CORS 필터
+                .addFilterBefore(jwtAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class)
                 .addFilterBefore(jwtAuthorizationFilter(), UsernamePasswordAuthenticationFilter.class);
+
 
         return http.build();
     }
