+++ src/main/java/com/circle/circle_backend/security/filter/JwtAuthorizationFilter.java
@@ -2,6 +2,7 @@ package com.circle.circle_backend.security.filter;
 
 import com.circle.circle_backend.security.utils.JwtTokenUtils;
 import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -13,6 +14,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
@@ -31,24 +33,26 @@ public class JwtAuthorizationFilter extends OncePerRequestFilter {
                                     HttpServletResponse response,
                                     FilterChain filterChain) throws ServletException, IOException {
 
-        String requestUri = request.getRequestURI();
-
-        // 인증이 필요 없는 경로
-        if (requestUri.startsWith("/api")) {
-            filterChain.doFilter(request, response);
-            return;
-        }
-
         String accessToken = jwtTokenUtils.getAccessToken(request); // 헤더에서 AccessToken 가져오기
 
-        if (jwtTokenUtils.validateToken(accessToken)) {
-            log.info("유효한 accessToken");
-            authenticateWithAccessToken(accessToken);
-        } else {
-            log.info("유효하지 않은 accessToken");
-            Claims claims = jwtTokenUtils.parseClaims(accessToken);
-            createNewAccessTokenWithRefreshToken(request, response, claims.getSubject());
+        try {
+            if (StringUtils.hasText(accessToken)) {
+                if (jwtTokenUtils.validateToken(accessToken)) {
+                    log.info("유효한 accessToken");
+                    authenticateWithAccessToken(accessToken);
+                } else {
+                    log.info("유효하지 않은 accessToken");
+                    Claims claims = jwtTokenUtils.parseClaims(accessToken);
+                    createNewAccessTokenWithRefreshToken(request, response, claims.getSubject());
+                }
+            }
+        } catch (ExpiredJwtException e) {
+            log.warn("Access Token has expired: {}", e.getMessage());
+        } catch (Exception e) {
+            log.error("Unexpected error during token validation: {}", e.getMessage());
+            SecurityContextHolder.clearContext(); // 인증 정보 초기화
         }
+
         filterChain.doFilter(request, response);
     }
 
@@ -74,6 +78,12 @@ public class JwtAuthorizationFilter extends OncePerRequestFilter {
                                                      HttpServletResponse response, String email) {
         log.info("refreshToken 검증 시도");
         String refreshToken = jwtTokenUtils.getRefreshToken(request);
+
+        if (refreshToken == null || refreshToken.isEmpty()) {
+            log.warn("Refresh Token is missing or empty. Clearing security context.");
+            SecurityContextHolder.clearContext();
+            return;
+        }
 
         // refreshToken이 유효한 토큰인지 확인
         if (jwtTokenUtils.validateToken(refreshToken)) {
