+++ src/main/java/com/circle/circle_backend/security/utils/JwtTokenUtils.java
@@ -51,7 +51,7 @@ public class JwtTokenUtils {
 
     // 헤더에서 accessToken 가져오기
     public String getAccessToken(HttpServletRequest request) {
-        String token = request.getHeader(BEARER_PREFIX);
+        String token = request.getHeader(AUTH_ACCESS_HEADER);
         if (!(StringUtils.hasText(token) && token.startsWith(BEARER_PREFIX))) {
             return null;
         }
